Cleanup token scripts
